rustc_hir

Module attrs

Data structures for representing parsed attributes in the Rust compiler. Formerly rustc_attr_data_structures.

For detailed documentation about attribute processing, see rustc_attr_parsing.

Re-exports

pub use ReprAttr::*;

Modules

data_structures 🔒

encode_cross_crate 🔒

pretty_printing 🔒

Structs

Deprecation

LinkEntry

SanitizerSet

StrippedCfgItem

Enums

AttributeKind

Represents parsed built-in inert attributes.

CfgEntry

CoverageAttrKind

Successfully-parsed value of a #[coverage(..)] attribute.

DeprecatedSince

Release in which an API is deprecated.

EncodeCrossCrate

InlineAttr

InstructionSetAttr

IntType

Linkage

Possible values for the #[linkage] attribute, allowing to specify the linkage type for a MonoItem.

MacroUseArgs

MirDialect

MirPhase

NativeLibKind

OptimizeAttr

PeImportNameType

Different ways that the PE Format can decorate a symbol name. From https://docs.microsoft.com/en-us/windows/win32/debug/pe-format#import-name-type

ReprAttr

TransparencyError

UsedBy

There are three valid forms of the attribute: #[used], which is semantically equivalent to #[used(linker)] except that the latter is currently unstable. #[used(compiler)] #[used(linker)]

Traits

PrintAttribute

This trait is used to print attributes in rustc_hir_pretty.