rustc_hir::attrs

Enum AttributeKind

pub enum AttributeKind {

    Align {
        align: Align,
        span: Span,
    },
    AllowConstFnUnstable(ThinVec<Symbol>, Span),
    AllowIncoherentImpl(Span),
    AllowInternalUnstable(ThinVec<(Symbol, Span)>, Span),
    AsPtr(Span),
    AutomaticallyDerived(Span),
    BodyStability {
        stability: DefaultBodyStability,
        span: Span,
    },
    CoherenceIsCore,
    Coinductive(Span),
    Cold(Span),
    Confusables {
        symbols: ThinVec<Symbol>,
        first_span: Span,
    },
    ConstContinue(Span),
    ConstStability {
        stability: PartialConstStability,
        span: Span,
    },
    ConstStabilityIndirect,
    ConstTrait(Span),
    Coroutine(Span),
    Coverage(Span, CoverageAttrKind),
    DenyExplicitImpl(Span),
    Deprecation {
        deprecation: Deprecation,
        span: Span,
    },
    DoNotImplementViaObject(Span),
    DocComment {
        style: AttrStyle,
        kind: CommentKind,
        span: Span,
        comment: Symbol,
    },
    Dummy,
    ExportName {
        name: Symbol,
        span: Span,
    },
    ExportStable,
    FfiConst(Span),
    FfiPure(Span),
    Fundamental,
    Ignore {
        span: Span,
        reason: Option<Symbol>,
    },
    Inline(InlineAttr, Span),
    LinkName {
        name: Symbol,
        span: Span,
    },
    LinkOrdinal {
        ordinal: u16,
        span: Span,
    },
    LinkSection {
        name: Symbol,
        span: Span,
    },
    LoopMatch(Span),
    MacroEscape(Span),
    MacroTransparency(Transparency),
    MacroUse {
        span: Span,
        arguments: MacroUseArgs,
    },
    Marker(Span),
    MayDangle(Span),
    MustUse {
        span: Span,
        reason: Option<Symbol>,
    },
    Naked(Span),
    NoImplicitPrelude(Span),
    NoMangle(Span),
    NonExhaustive(Span),
    Optimize(OptimizeAttr, Span),
    ParenSugar(Span),
    PassByValue(Span),
    Path(Symbol, Span),
    Pointee(Span),
    ProcMacro(Span),
    ProcMacroAttribute(Span),
    ProcMacroDerive {
        trait_name: Symbol,
        helper_attrs: ThinVec<Symbol>,
        span: Span,
    },
    PubTransparent(Span),
    Repr {
        reprs: ThinVec<(ReprAttr, Span)>,
        first_span: Span,
    },
    RustcBuiltinMacro {
        builtin_name: Option<Symbol>,
        helper_attrs: ThinVec<Symbol>,
        span: Span,
    },
    RustcLayoutScalarValidRangeEnd(Box<u128>, Span),
    RustcLayoutScalarValidRangeStart(Box<u128>, Span),
    RustcObjectLifetimeDefault,
    ShouldPanic {
        reason: Option<Symbol>,
        span: Span,
    },
    SkipDuringMethodDispatch {
        array: bool,
        boxed_slice: bool,
        span: Span,
    },
    SpecializationTrait(Span),
    Stability {
        stability: Stability,
        span: Span,
    },
    StdInternalSymbol(Span),
    TargetFeature(ThinVec<(Symbol, Span)>, Span),
    TrackCaller(Span),
    TypeConst(Span),
    UnsafeSpecializationMarker(Span),
    UnstableFeatureBound(ThinVec<(Symbol, Span)>),
    Used {
        used_by: UsedBy,
        span: Span,
    },

}

Represents parsed built-in inert attributes.

Overview

These attributes are markers that guide the compilation process and are never expanded into other code. They persist throughout the compilation phases, from AST to HIR and beyond.

Attribute Processing

While attributes are initially parsed by rustc_parse into ast::Attribute, they still contain raw token streams because different attributes have different internal structures. This enum represents the final, fully parsed form of these attributes, where each variant contains all the information and structure relevant for the specific attribute.

Some attributes can be applied multiple times to the same item, and they are “collapsed” into a single semantic attribute. For example:

#[repr(C)]
#[repr(packed)]
struct S { }

This is equivalent to #[repr(C, packed)] and results in a single AttributeKind::Repr containing both C and packed annotations. This collapsing happens during parsing and is reflected in the data structures defined in this enum.

Usage

These parsed attributes are used throughout the compiler to:

Control code generation (e.g., #[repr])

Mark API stability (#[stable], #[unstable])

Provide documentation (#[doc])

Guide compiler behavior (e.g., #[allow_internal_unstable])

Note on Attribute Organization

Some attributes like InlineAttr, OptimizeAttr, and InstructionSetAttr are defined separately from this enum because they are used in specific compiler phases (like code generation) and don’t need to persist throughout the entire compilation process. They are typically processed and converted into their final form earlier in the compilation pipeline.

For example:

InlineAttr is used during code generation to control function inlining

OptimizeAttr is used to control optimization levels

InstructionSetAttr is used for target-specific code generation

These attributes are handled by their respective compiler passes in the rustc_codegen_ssa crate and don’t need to be preserved in the same way as the attributes in this enum.

For more details on attribute parsing, see the rustc_attr_parsing crate.

Variants

Align

Represents #[align(N)].

Fields

align: Align

span: Span

AllowConstFnUnstable(ThinVec<Symbol>, Span)

Represents #[rustc_allow_const_fn_unstable].

AllowIncoherentImpl(Span)

Represents #[rustc_allow_incoherent_impl].

AllowInternalUnstable(ThinVec<(Symbol, Span)>, Span)

Represents #[allow_internal_unstable].

AsPtr(Span)

Represents #[rustc_as_ptr] (used by the dangling_pointers_from_temporaries lint).

AutomaticallyDerived(Span)

Represents #[automatically_derived]

BodyStability

Represents #[rustc_default_body_unstable].

Fields

stability: DefaultBodyStability

span: Span

Span of the #[rustc_default_body_unstable(...)] attribute

CoherenceIsCore

Represents #[rustc_coherence_is_core].

Coinductive(Span)

Represents #[rustc_coinductive].

Cold(Span)

Represents #[cold].

Confusables

Represents #[rustc_confusables].

Fields

symbols: ThinVec<Symbol>

first_span: Span

ConstContinue(Span)

Represents #[const_continue].

ConstStability

Represents #[rustc_const_stable] and #[rustc_const_unstable].

Fields

stability: PartialConstStability

span: Span

Span of the #[rustc_const_stable(...)] or #[rustc_const_unstable(...)] attribute

ConstStabilityIndirect

Represents #[rustc_const_stable_indirect].

ConstTrait(Span)

Represents #[const_trait].

Coroutine(Span)

Represents #[coroutine].

Coverage(Span, CoverageAttrKind)

Represents #[coverage(..)].

DenyExplicitImpl(Span)

Represents #[rustc_deny_explicit_impl].

Deprecation

Represents #[deprecated].

Fields

deprecation: Deprecation

span: Span

DoNotImplementViaObject(Span)

Represents #[rustc_do_not_implement_via_object].

DocComment

Represents #[doc].

Fields

style: AttrStyle

kind: CommentKind

span: Span

comment: Symbol

Dummy

Represents #[rustc_dummy].

ExportName

Represents #[export_name].

Fields

name: Symbol

The name to export this item with. It may not contain \0 bytes as it will be converted to a null-terminated string.

span: Span

ExportStable

Represents #[export_stable].

FfiConst(Span)

Represents #[ffi_const].

FfiPure(Span)

Represents #[ffi_pure].

Fundamental

Represents #[fundamental].

Ignore

Represents #[ignore]

Fields

span: Span

reason: Option<Symbol>

ignore can optionally have a reason: #[ignore = "reason this is ignored"]

Inline(InlineAttr, Span)

Represents #[inline] and #[rustc_force_inline].

LinkName

Represents #[link_name].

Fields

name: Symbol

span: Span

LinkOrdinal

Represents #[link_ordinal].

Fields

ordinal: u16

span: Span

LinkSection

Represents #[link_section]

Fields

name: Symbol

span: Span

LoopMatch(Span)

Represents #[loop_match].

MacroEscape(Span)

Represents #[macro_escape].

MacroTransparency(Transparency)

Represents #[rustc_macro_transparency].

MacroUse

Represents #[macro_use].

Fields

span: Span

arguments: MacroUseArgs

Marker(Span)

Represents #[marker].

MayDangle(Span)

Represents #[may_dangle].

MustUse

Represents #[must_use].

Fields

span: Span

reason: Option<Symbol>

must_use can optionally have a reason: #[must_use = "reason this must be used"]

Naked(Span)

Represents #[naked]

NoImplicitPrelude(Span)

Represents #[no_implicit_prelude]

NoMangle(Span)

Represents #[no_mangle]

NonExhaustive(Span)

Represents #[non_exhaustive]

Optimize(OptimizeAttr, Span)

Represents #[optimize(size|speed)]

ParenSugar(Span)

Represents #[rustc_paren_sugar].

PassByValue(Span)

Represents #[rustc_pass_by_value] (used by the rustc_pass_by_value lint).

Path(Symbol, Span)

Represents #[path]

Pointee(Span)

Represents #[pointee]

ProcMacro(Span)

Represents #[proc_macro]

ProcMacroAttribute(Span)

Represents #[proc_macro_attribute]

ProcMacroDerive

Represents #[proc_macro_derive]

Fields

trait_name: Symbol

helper_attrs: ThinVec<Symbol>

span: Span

PubTransparent(Span)

Represents #[rustc_pub_transparent] (used by the repr_transparent_external_private_fields lint).

Repr

Represents #[repr].

Fields

reprs: ThinVec<(ReprAttr, Span)>

first_span: Span

RustcBuiltinMacro

Represents #[rustc_builtin_macro].

Fields

builtin_name: Option<Symbol>

helper_attrs: ThinVec<Symbol>

span: Span

RustcLayoutScalarValidRangeEnd(Box<u128>, Span)

Represents #[rustc_layout_scalar_valid_range_end].

RustcLayoutScalarValidRangeStart(Box<u128>, Span)

Represents #[rustc_layout_scalar_valid_range_start].

RustcObjectLifetimeDefault

Represents #[rustc_object_lifetime_default].

ShouldPanic

Represents #[should_panic]

Fields

reason: Option<Symbol>

span: Span

SkipDuringMethodDispatch

Represents #[rustc_skip_during_method_dispatch].

Fields

array: bool

boxed_slice: bool

span: Span

SpecializationTrait(Span)

Represents #[rustc_specialization_trait].

Stability

Represents #[stable], #[unstable] and #[rustc_allowed_through_unstable_modules].

Fields

stability: Stability

span: Span

Span of the attribute.

StdInternalSymbol(Span)

Represents #[rustc_std_internal_symbol].

TargetFeature(ThinVec<(Symbol, Span)>, Span)

Represents #[target_feature(enable = "...")]

TrackCaller(Span)

Represents #[track_caller]

TypeConst(Span)

Represents #[type_const].

UnsafeSpecializationMarker(Span)

Represents #[rustc_unsafe_specialization_marker].

UnstableFeatureBound(ThinVec<(Symbol, Span)>)

Represents #[unstable_feature_bound].

Used

Represents #[used]

Fields

used_by: UsedBy

span: Span

Implementations

impl AttributeKind

pub fn encode_cross_crate(&self) -> EncodeCrossCrate