Enum rustc_hir::def::DefKind

pub enum DefKind {

    Mod,
    Struct,
    Union,
    Enum,
    Variant,
    Trait,
    TyAlias {
        lazy: bool,
    },
    ForeignTy,
    TraitAlias,
    AssocTy,
    TyParam,
    Fn,
    Const,
    ConstParam,
    Static(Mutability),
    Ctor(CtorOf, CtorKind),
    AssocFn,
    AssocConst,
    Macro(MacroKind),
    ExternCrate,
    Use,
    ForeignMod,
    AnonConst,
    InlineConst,
    OpaqueTy,
    Field,
    LifetimeParam,
    GlobalAsm,
    Impl {
        of_trait: bool,
    },
    Closure,
    Generator,

}

What kind of definition something is; e.g., mod vs struct.

Variants

Mod

Struct

Refers to the struct itself, DefKind::Ctor refers to its constructor if it exists.

Union

Enum

Variant

Refers to the variant itself, DefKind::Ctor refers to its constructor if it exists.

Trait

TyAlias

Fields

lazy: bool

Type alias: type Foo = Bar;

ForeignTy

Type from an extern block.

TraitAlias

Trait alias: trait IntIterator = Iterator<Item = i32>;

AssocTy

Associated type: trait MyTrait { type Assoc; }

TyParam

Type parameter: the T in struct Vec<T> { ... }

Fn

Const

ConstParam

Constant generic parameter: struct Foo<const N: usize> { ... }

Static(Mutability)

Ctor(CtorOf, CtorKind)

Refers to the struct or enum variant’s constructor.

The reason Ctor exists in addition to DefKind::Struct and DefKind::Variant is because structs and enum variants exist in the type namespace, whereas struct and enum variant constructors exist in the value namespace.

You may wonder why enum variants exist in the type namespace as opposed to the value namespace. Check out RFC 2593 for intuition on why that is.

AssocFn

Associated function: impl MyStruct { fn associated() {} } or trait Foo { fn associated() {} }

AssocConst

Associated constant: trait MyTrait { const ASSOC: usize; }

Macro(MacroKind)

ExternCrate

Use

ForeignMod

An extern block.

AnonConst

Anonymous constant, e.g. the 1 + 2 in [u8; 1 + 2]

InlineConst

An inline constant, e.g. const { 1 + 2 }

OpaqueTy

Opaque type, aka impl Trait.

Field

LifetimeParam

Lifetime parameter: the 'a in struct Foo<'a> { ... }

GlobalAsm

A use of global_asm!.

Impl

Fields

of_trait: bool

Closure

Generator

Implementations

impl DefKind

pub fn descr(self, def_id: DefId) -> &'static str

Get an English description for the item’s kind.

If you have access to TyCtxt, use TyCtxt::def_descr or TyCtxt::def_kind_descr instead, because they give better information for generators and associated functions.

pub fn article(&self) -> &'static str

Gets an English article for the definition.

If you have access to TyCtxt, use TyCtxt::def_descr_article or TyCtxt::def_kind_descr_article instead, because they give better information for generators and associated functions.

pub fn ns(&self) -> Option<Namespace>

pub fn is_fn_like(self) -> bool

pub fn has_codegen_attrs(self) -> bool

Whether query get_codegen_attrs should be used with this definition.