Struct rustc_hir::def::PerNS

pub struct PerNS<T> {
    pub value_ns: T,
    pub type_ns: T,
    pub macro_ns: T,
}

Just a helper ‒ separate structure for each namespace.

Fields
value_ns: Ttype_ns: Tmacro_ns: T
Implementations

impl<T> PerNS<T>

pub fn map<U, F: FnMut(T) -> U>(self, f: F) -> PerNS<U>

pub fn into_iter(self) -> IntoIter<T, 3>

pub fn iter(&self) -> IntoIter<&T, 3>

impl<T> PerNS<Option<T>>

pub fn is_empty(&self) -> bool

Returns true if all the items in this collection are None.

pub fn present_items(self) -> impl Iterator<Item = T>

Returns an iterator over the items which are Some.