Struct rustc_hir::definitions::Definitions

pub struct Definitions {
    table: DefPathTable,
    next_disambiguator: FxHashMap<(LocalDefId, DefPathData), u32>,
    stable_crate_id: StableCrateId,
}

The definition table containing node definitions. It holds the DefPathTable for LocalDefIds/DefPaths. It also stores mappings to convert LocalDefIds to/from HirIds.

Fields
table: DefPathTablenext_disambiguator: FxHashMap<(LocalDefId, DefPathData), u32>stable_crate_id: StableCrateId

The StableCrateId of the local crate.

Implementations

impl Definitions

pub fn def_path_table(&self) -> &DefPathTable

pub fn def_index_count(&self) -> usize

Gets the number of definitions.

pub fn def_key(&self, id: LocalDefId) -> DefKey

pub fn def_path_hash(&self, id: LocalDefId) -> DefPathHash

pub fn def_path(&self, id: LocalDefId) -> DefPath

Returns the path from the crate root to index. The root nodes are not included in the path (i.e., this will be an empty vector for the crate root). For an inlined item, this will be the path of the item in the external crate (but the path will begin with the path to the external crate).

pub fn new(stable_crate_id: StableCrateId) -> Definitions

Adds a root definition (no parent) and a few other reserved definitions.

pub fn create_def( &mut self, parent: LocalDefId, data: DefPathData ) -> LocalDefId

Adds a definition with a parent definition.

pub fn local_def_path_hash_to_def_id( &self, hash: DefPathHash, err: &mut dyn FnMut() -> ! ) -> LocalDefId

pub fn def_path_hash_to_def_index_map(&self) -> &DefPathHashMap

pub fn num_definitions(&self) -> usize