Enum rustc_hir::ExprKind

pub enum ExprKind<'hir> {

    Box(&'hir Expr<'hir>),
    ConstBlock(AnonConst),
    Array(&'hir [Expr<'hir>]),
    Call(&'hir Expr<'hir>, &'hir [Expr<'hir>]),
    MethodCall(&'hir PathSegment<'hir>, &'hir Expr<'hir>, &'hir [Expr<'hir>], Span),
    Tup(&'hir [Expr<'hir>]),
    Binary(BinOp, &'hir Expr<'hir>, &'hir Expr<'hir>),
    Unary(UnOp, &'hir Expr<'hir>),
    Lit(Lit),
    Cast(&'hir Expr<'hir>, &'hir Ty<'hir>),
    Type(&'hir Expr<'hir>, &'hir Ty<'hir>),
    DropTemps(&'hir Expr<'hir>),
    Let(&'hir Let<'hir>),
    If(&'hir Expr<'hir>, &'hir Expr<'hir>, Option<&'hir Expr<'hir>>),
    Loop(&'hir Block<'hir>, Option<Label>, LoopSource, Span),
    Match(&'hir Expr<'hir>, &'hir [Arm<'hir>], MatchSource),
    Closure(&'hir Closure<'hir>),
    Block(&'hir Block<'hir>, Option<Label>),
    Assign(&'hir Expr<'hir>, &'hir Expr<'hir>, Span),
    AssignOp(BinOp, &'hir Expr<'hir>, &'hir Expr<'hir>),
    Field(&'hir Expr<'hir>, Ident),
    Index(&'hir Expr<'hir>, &'hir Expr<'hir>),
    Path(QPath<'hir>),
    AddrOf(BorrowKind, Mutability, &'hir Expr<'hir>),
    Break(Destination, Option<&'hir Expr<'hir>>),
    Continue(Destination),
    Ret(Option<&'hir Expr<'hir>>),
    InlineAsm(&'hir InlineAsm<'hir>),
    Struct(&'hir QPath<'hir>, &'hir [ExprField<'hir>], Option<&'hir Expr<'hir>>),
    Repeat(&'hir Expr<'hir>, ArrayLen),
    Yield(&'hir Expr<'hir>, YieldSource),
    Err,

}

Variants

Box(&'hir Expr<'hir>)

A box x expression.

ConstBlock(AnonConst)

Allow anonymous constants from an inline const block

Array(&'hir [Expr<'hir>])

An array (e.g., [a, b, c, d]).

Call(&'hir Expr<'hir>, &'hir [Expr<'hir>])

A function call.

The first field resolves to the function itself (usually an ExprKind::Path), and the second field is the list of arguments. This also represents calling the constructor of tuple-like ADTs such as tuple structs and enum variants.

MethodCall(&'hir PathSegment<'hir>, &'hir Expr<'hir>, &'hir [Expr<'hir>], Span)

A method call (e.g., x.foo::<'static, Bar, Baz>(a, b, c, d)).

The PathSegment represents the method name and its generic arguments (within the angle brackets). The &Expr is the expression that evaluates to the object on which the method is being called on (the receiver), and the &[Expr] is the rest of the arguments. Thus, x.foo::<Bar, Baz>(a, b, c, d) is represented as ExprKind::MethodCall(PathSegment { foo, [Bar, Baz] }, x, [a, b, c, d], span). The final Span represents the span of the function and arguments (e.g. foo::<Bar, Baz>(a, b, c, d) in x.foo::<Bar, Baz>(a, b, c, d)

To resolve the called method to a DefId, call type_dependent_def_id with the hir_id of the MethodCall node itself.

Tup(&'hir [Expr<'hir>])

A tuple (e.g., (a, b, c, d)).

Binary(BinOp, &'hir Expr<'hir>, &'hir Expr<'hir>)

A binary operation (e.g., a + b, a * b).

Unary(UnOp, &'hir Expr<'hir>)

A unary operation (e.g., !x, *x).

Lit(Lit)

A literal (e.g., 1, "foo").

Cast(&'hir Expr<'hir>, &'hir Ty<'hir>)

A cast (e.g., foo as f64).

Type(&'hir Expr<'hir>, &'hir Ty<'hir>)

A type reference (e.g., Foo).

DropTemps(&'hir Expr<'hir>)

Wraps the expression in a terminating scope. This makes it semantically equivalent to { let _t = expr; _t }.

This construct only exists to tweak the drop order in HIR lowering. An example of that is the desugaring of for loops.

Let(&'hir Let<'hir>)

A let $pat = $expr expression.

These are not Local and only occur as expressions. The let Some(x) = foo() in if let Some(x) = foo() is an example of Let(..).

If(&'hir Expr<'hir>, &'hir Expr<'hir>, Option<&'hir Expr<'hir>>)

An if block, with an optional else block.

I.e., if <expr> { <expr> } else { <expr> }.

Loop(&'hir Block<'hir>, Option<Label>, LoopSource, Span)

A conditionless loop (can be exited with break, continue, or return).

I.e., 'label: loop { <block> }.

The Span is the loop header (for x in y/while let pat = expr).

Match(&'hir Expr<'hir>, &'hir [Arm<'hir>], MatchSource)

A match block, with a source that indicates whether or not it is the result of a desugaring, and if so, which kind.

Closure(&'hir Closure<'hir>)

A closure (e.g., move |a, b, c| {a + b + c}).

The Span is the argument block |...|.

This may also be a generator literal or an async block as indicated by the Option<Movability>.

Block(&'hir Block<'hir>, Option<Label>)

A block (e.g., 'label: { ... }).

Assign(&'hir Expr<'hir>, &'hir Expr<'hir>, Span)

An assignment (e.g., a = foo()).

AssignOp(BinOp, &'hir Expr<'hir>, &'hir Expr<'hir>)

An assignment with an operator.

E.g., a += 1.

Field(&'hir Expr<'hir>, Ident)

Access of a named (e.g., obj.foo) or unnamed (e.g., obj.0) struct or tuple field.

Index(&'hir Expr<'hir>, &'hir Expr<'hir>)

An indexing operation (foo[2]).

Path(QPath<'hir>)

Path to a definition, possibly containing lifetime or type parameters.

AddrOf(BorrowKind, Mutability, &'hir Expr<'hir>)

A referencing operation (i.e., &a or &mut a).

Break(Destination, Option<&'hir Expr<'hir>>)

A break, with an optional label to break.

Continue(Destination)

A continue, with an optional label.

Ret(Option<&'hir Expr<'hir>>)

A return, with an optional value to be returned.

InlineAsm(&'hir InlineAsm<'hir>)

Inline assembly (from asm!), with its outputs and inputs.

Struct(&'hir QPath<'hir>, &'hir [ExprField<'hir>], Option<&'hir Expr<'hir>>)

A struct or struct-like variant literal expression.

E.g., Foo {x: 1, y: 2}, or Foo {x: 1, .. base}, where base is the Option<Expr>.

Repeat(&'hir Expr<'hir>, ArrayLen)

An array literal constructed from one repeated element.

E.g., [1; 5]. The first expression is the element to be repeated; the second is the number of times to repeat it.

Yield(&'hir Expr<'hir>, YieldSource)

A suspension point for generators (i.e., yield <expr>).

Err

A placeholder for an expression that wasn’t syntactically well formed in some way.