Enum rustc_hir::Node

pub enum Node<'hir> {

    Param(&'hir Param<'hir>),
    Item(&'hir Item<'hir>),
    ForeignItem(&'hir ForeignItem<'hir>),
    TraitItem(&'hir TraitItem<'hir>),
    ImplItem(&'hir ImplItem<'hir>),
    Variant(&'hir Variant<'hir>),
    Field(&'hir FieldDef<'hir>),
    AnonConst(&'hir AnonConst),
    Expr(&'hir Expr<'hir>),
    ExprField(&'hir ExprField<'hir>),
    Stmt(&'hir Stmt<'hir>),
    PathSegment(&'hir PathSegment<'hir>),
    Ty(&'hir Ty<'hir>),
    TypeBinding(&'hir TypeBinding<'hir>),
    TraitRef(&'hir TraitRef<'hir>),
    Pat(&'hir Pat<'hir>),
    PatField(&'hir PatField<'hir>),
    Arm(&'hir Arm<'hir>),
    Block(&'hir Block<'hir>),
    Local(&'hir Local<'hir>),
    Ctor(&'hir VariantData<'hir>),
    Lifetime(&'hir Lifetime),
    GenericParam(&'hir GenericParam<'hir>),
    Crate(&'hir Mod<'hir>),
    Infer(&'hir InferArg),

}

Variants

Param(&'hir Param<'hir>)

Item(&'hir Item<'hir>)

ForeignItem(&'hir ForeignItem<'hir>)

TraitItem(&'hir TraitItem<'hir>)

ImplItem(&'hir ImplItem<'hir>)

Variant(&'hir Variant<'hir>)

Field(&'hir FieldDef<'hir>)

AnonConst(&'hir AnonConst)

Expr(&'hir Expr<'hir>)

ExprField(&'hir ExprField<'hir>)

Stmt(&'hir Stmt<'hir>)

PathSegment(&'hir PathSegment<'hir>)

Ty(&'hir Ty<'hir>)

TypeBinding(&'hir TypeBinding<'hir>)

TraitRef(&'hir TraitRef<'hir>)

Pat(&'hir Pat<'hir>)

PatField(&'hir PatField<'hir>)

Arm(&'hir Arm<'hir>)

Block(&'hir Block<'hir>)

Local(&'hir Local<'hir>)

Ctor(&'hir VariantData<'hir>)

Ctor refers to the constructor of an enum variant or struct. Only tuple or unit variants with synthesized constructors.

Lifetime(&'hir Lifetime)

GenericParam(&'hir GenericParam<'hir>)

Crate(&'hir Mod<'hir>)

Infer(&'hir InferArg)

Implementations

impl<'hir> Node<'hir>

pub fn ident(&self) -> Option<Ident>

Get the identifier of this Node, if applicable.

Edge cases

Calling .ident() on a Node::Ctor will return None because Ctors do not have identifiers themselves. Instead, call .ident() on the parent struct/variant, like so:

ctor
    .ctor_hir_id()
    .and_then(|ctor_id| tcx.hir().find(tcx.hir().get_parent_node(ctor_id)))
    .and_then(|parent| parent.ident())

pub fn fn_decl(self) -> Option<&'hir FnDecl<'hir>>

pub fn fn_sig(self) -> Option<&'hir FnSig<'hir>>

pub fn body_id(&self) -> Option<BodyId>

pub fn generics(self) -> Option<&'hir Generics<'hir>>

pub fn as_owner(self) -> Option<OwnerNode<'hir>>

pub fn fn_kind(self) -> Option<FnKind<'hir>>

pub fn tuple_fields(&self) -> Option<&'hir [FieldDef<'hir>]>

Get the fields for the tuple-constructor, if this node is a tuple constructor, otherwise None