Enum rustc_hir::TyKind

pub enum TyKind<'hir> {

    Slice(&'hir Ty<'hir>),
    Array(&'hir Ty<'hir>, ArrayLen),
    Ptr(MutTy<'hir>),
    Rptr(&'hir Lifetime, MutTy<'hir>),
    BareFn(&'hir BareFnTy<'hir>),
    Never,
    Tup(&'hir [Ty<'hir>]),
    Path(QPath<'hir>),
    OpaqueDef(ItemId, &'hir [GenericArg<'hir>], bool),
    TraitObject(&'hir [PolyTraitRef<'hir>], &'hir Lifetime, TraitObjectSyntax),
    Typeof(AnonConst),
    Infer,
    Err,

}

The various kinds of types recognized by the compiler.

Variants

Slice(&'hir Ty<'hir>)

A variable length slice (i.e., [T]).

Array(&'hir Ty<'hir>, ArrayLen)

A fixed length array (i.e., [T; n]).

Ptr(MutTy<'hir>)

A raw pointer (i.e., *const T or *mut T).

Rptr(&'hir Lifetime, MutTy<'hir>)

A reference (i.e., &'a T or &'a mut T).

BareFn(&'hir BareFnTy<'hir>)

A bare function (e.g., fn(usize) -> bool).

Never

The never type (!).

Tup(&'hir [Ty<'hir>])

A tuple ((A, B, C, D, ...)).

Path(QPath<'hir>)

A path to a type definition (module::module::...::Type), or an associated type (e.g., <Vec<T> as Trait>::Type or <T>::Target).

Type parameters may be stored in each PathSegment.

OpaqueDef(ItemId, &'hir [GenericArg<'hir>], bool)

An opaque type definition itself. This is only used for impl Trait.

The generic argument list contains the lifetimes (and in the future possibly parameters) that are actually bound on the impl Trait.

The last parameter specifies whether this opaque appears in a trait definition.

TraitObject(&'hir [PolyTraitRef<'hir>], &'hir Lifetime, TraitObjectSyntax)

A trait object type Bound1 + Bound2 + Bound3 where Bound is a trait or a lifetime.

Typeof(AnonConst)

Unused for now.

Infer

TyKind::Infer means the type should be inferred instead of it having been specified. This can appear anywhere in a type.

Err

Placeholder for a type that has failed to be defined.