rustc_hir::hir

Enum Attribute

pub enum Attribute {
    Parsed(AttributeKind),
    Unparsed(Box<AttrItem>),
}

Variants

Parsed(AttributeKind)

A parsed built-in attribute.

Each attribute has a span connected to it. However, you must be somewhat careful using it. That’s because sometimes we merge multiple attributes together, like when an item has multiple repr attributes. In this case the span might not be very useful.

Unparsed(Box<AttrItem>)

An attribute that could not be parsed, out of a token-like representation. This is the case for custom tool attributes.

Implementations

impl Attribute

pub fn get_normal_item(&self) -> &AttrItem

pub fn unwrap_normal_item(self) -> AttrItem

pub fn value_lit(&self) -> Option<&MetaItemLit>

pub fn is_parsed_attr(&self) -> bool

impl Attribute

pub fn id(&self) -> AttrId

pub fn name(&self) -> Option<Symbol>

pub fn meta_item_list(&self) -> Option<ThinVec<MetaItemInner>>

pub fn value_str(&self) -> Option<Symbol>

pub fn value_span(&self) -> Option<Span>

pub fn ident(&self) -> Option<Ident>

pub fn path_matches(&self, name: &[Symbol]) -> bool

pub fn is_doc_comment(&self) -> bool

pub fn has_name(&self, name: Symbol) -> bool

pub fn has_any_name(&self, names: &[Symbol]) -> bool

pub fn span(&self) -> Span

pub fn is_word(&self) -> bool

pub fn path(&self) -> SmallVec<[Symbol; 1]>

pub fn ident_path(&self) -> Option<SmallVec<[Ident; 1]>>

pub fn doc_str(&self) -> Option<Symbol>

pub fn is_proc_macro_attr(&self) -> bool

pub fn doc_str_and_comment_kind(&self) -> Option<(Symbol, CommentKind)>