Enum rustc_hir::hir::LifetimeName

pub enum LifetimeName {
    Param(LocalDefId, ParamName),
    ImplicitObjectLifetimeDefault,
    Error,
    Infer,
    Static,
}

Variants

Param(LocalDefId, ParamName)

User-given names or fresh (synthetic) names.

ImplicitObjectLifetimeDefault

Implicit lifetime in a context like dyn Foo. This is distinguished from implicit lifetimes elsewhere because the lifetime that they default to must appear elsewhere within the enclosing type. This means that, in an impl Trait context, we don’t have to create a parameter for them. That is, impl Trait<Item = &u32> expands to an opaque type like type Foo<'a> = impl Trait<Item = &'a u32>, but impl Trait<item = dyn Bar> expands to type Foo = impl Trait<Item = dyn Bar + 'static>. The latter uses ImplicitObjectLifetimeDefault so that surrounding code knows not to create a lifetime parameter.

Error

Indicates an error during lowering (usually '_ in wrong place) that was already reported.

Infer

User wrote an anonymous lifetime, either '_ or nothing. The semantics of this lifetime should be inferred by typechecking code.

Static

User wrote 'static.

Implementations

impl LifetimeName

pub fn ident(&self) -> Ident

pub fn is_anonymous(&self) -> bool

pub fn is_elided(&self) -> bool

fn is_static(&self) -> bool

pub fn normalize_to_macros_2_0(&self) -> LifetimeName