Module rustc_hir::hir_id

Structs

HirId

Uniquely identifies a node in the HIR of the current crate. It is composed of the owner, which is the LocalDefId of the directly enclosing hir::Item, hir::TraitItem, or hir::ImplItem (i.e., the closest “item-like”), and the local_id which is unique within the given owner.

ItemLocalId

An ItemLocalId uniquely identifies something within a given “item-like”; that is, within a hir::Item, hir::TraitItem, or hir::ImplItem. There is no guarantee that the numerical value of a given ItemLocalId corresponds to the node’s position within the owning item in any way, but there is a guarantee that the LocalItemIds within an owner occupy a dense range of integers starting at zero, so a mapping that maps all or most nodes within an “item-like” to something else can be implemented by a Vec instead of a tree or hash map.

OwnerId

Constants

CRATE_HIR_ID

The HirId corresponding to CRATE_NODE_ID and CRATE_DEF_ID.

CRATE_OWNER_ID

Type Definitions

HirIdMap

HirIdMapEntry

HirIdSet

ItemLocalMap

ItemLocalMapEntry

ItemLocalSet