Trait rustc_hir::intravisit::nested_filter::NestedFilter

pub trait NestedFilter<'hir> {
    type Map: Map<'hir>;

    const INTER: bool;
    const INTRA: bool;
}

Specifies what nested things a visitor wants to visit. By “nested things”, we are referring to bits of HIR that are not directly embedded within one another but rather indirectly, through a table in the crate. This is done to control dependencies during incremental compilation: the non-inline bits of HIR can be tracked and hashed separately.

The most common choice is OnlyBodies, which will cause the visitor to visit fn bodies for fns that it encounters, and closure bodies, but skip over nested item-like things.

See the comments on ItemLikeVisitor for more details on the overall visit strategy.

Required Associated Types

type Map: Map<'hir>

Required Associated Constants

const INTER: bool

Whether the visitor visits nested “item-like” things. E.g., item, impl-item.

const INTRA: bool

Whether the visitor visits “intra item-like” things. E.g., function body, closure, AnonConst