rustc_hir::intravisit

Trait VisitorExt

pub trait VisitorExt<'v>: Visitor<'v> {
    // Provided methods
    fn visit_ty_unambig(&mut self, t: &'v Ty<'v>) -> Self::Result { ... }
    fn visit_const_arg_unambig(&mut self, c: &'v ConstArg<'v>) -> Self::Result { ... }
}

Provided Methods

fn visit_ty_unambig(&mut self, t: &'v Ty<'v>) -> Self::Result

Extension trait method to visit types in unambiguous positions, this is not directly on the Visitor trait as this method should never be overridden.

Named visit_ty_unambig instead of visit_unambig_ty to aid in discovery by IDes when v.visit_ty is written.

fn visit_const_arg_unambig(&mut self, c: &'v ConstArg<'v>) -> Self::Result

Extension trait method to visit consts in unambiguous positions, this is not directly on the Visitor trait as this method should never be overridden.

Named visit_const_arg_unambig instead of visit_unambig_const_arg to aid in discovery by IDes when v.visit_const_arg is written.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.