rustc_hir

Module stability

Structs

ConstStability

Represents the #[rustc_const_unstable] and #[rustc_const_stable] attributes.

DefaultBodyStability

Represents the #[rustc_default_body_unstable] attribute.

PartialConstStability

Excludes const_stable_indirect. This is necessary because when -Zforce-unstable-if-unmarked is set, we need to encode standalone #[rustc_const_stable_indirect] attributes

Stability

Represents the following attributes:

Enums

StabilityLevel

The available stability levels.

StableSince

Rust release in which a feature is stabilized.

UnstableReason

Constants

VERSION_PLACEHOLDER

The version placeholder that recently stabilized features contain inside the since field of the #[stable] attribute.