Trait rustc_hir::HashStableContext

pub trait HashStableContext: HashStableContext + HashStableContext {
    fn hash_body_id(&mut self, _: BodyId, hasher: &mut StableHasher);
}

Requirements for a StableHashingContext to be used in this crate. This is a hack to allow using the HashStable_Generic derive macro instead of implementing everything in rustc_middle.

Required Methods

fn hash_body_id(&mut self, _: BodyId, hasher: &mut StableHasher)