rustc_hir

Module def

Structs

MacroKinds

A set of macro kinds, for macros that can have more than one kind

PartialRes

The result of resolving a path before lowering to HIR, with “module” segments resolved and associated item segments deferred to type checking. base_res is the resolution of the resolved part of the path, unresolved_segments is the number of unresolved segments.

PerNS

Just a helper ‒ separate structure for each namespace.

Enums

CtorKind

What kind of constructor something is.

CtorOf

Encodes if a DefKind::Ctor is the constructor of an enum variant or a struct.

DefKind

What kind of definition something is; e.g., mod vs struct. enum DefPathData may need to be updated if a new variant is added here.

LifetimeRes

Resolution for a lifetime appearing in a type.

Namespace

Different kinds of symbols can coexist even if they share the same textual name. Therefore, they each have a separate universe (known as a “namespace”).

NonMacroAttrKind

An attribute that is not a macro; e.g., #[inline] or #[rustfmt::skip].

Res

The resolution of a path or export.

Type Aliases

DocLinkResMap