Enum rustc_hir::def::Res

pub enum Res<Id = HirId> {
    Def(DefKind, DefId),
    PrimTy(PrimTy),
    SelfTyParam {
        trait_: DefId,
    },
    SelfTyAlias {
        alias_to: DefId,
        forbid_generic: bool,
        is_trait_impl: bool,
    },
    SelfCtor(DefId),
    Local(Id),
    ToolMod,
    NonMacroAttr(NonMacroAttrKind),
    Err,
}

The resolution of a path or export.

For every path or identifier in Rust, the compiler must determine what the path refers to. This process is called name resolution, and Res is the primary result of name resolution.

For example, everything prefixed with /* Res */ in this example has an associated Res:

fn str_to_string(s: & /* Res */ str) -> /* Res */ String {
    /* Res */ String::from(/* Res */ s)
}

/* Res */ str_to_string("hello");

The associated Ress will be:

str will resolve to Res::PrimTy;

String will resolve to Res::Def, and the Res will include the DefId for String as defined in the standard library;

String::from will also resolve to Res::Def, with the DefId pointing to String::from;

s will resolve to Res::Local;

the call to str_to_string will resolve to Res::Def, with the DefId pointing to the definition of str_to_string in the current crate.

Variants

Def(DefKind, DefId)

Definition having a unique ID (DefId), corresponds to something defined in user code.

Not bound to a specific namespace.

PrimTy(PrimTy)

A primitive type such as i32 or str.

Belongs to the type namespace.

SelfTyParam

Fields

trait_: DefId

The trait this Self is a generic parameter for.

The Self type, as used within a trait.

Belongs to the type namespace.

See the examples on Res::SelfTyAlias for details.

SelfTyAlias

Fields

alias_to: DefId

The item introducing the Self type alias. Can be used in the type_of query to get the underlying type.

forbid_generic: bool

Whether the Self type is disallowed from mentioning generics (i.e. when used in an anonymous constant).

HACK(min_const_generics): self types also have an optional requirement to not mention any generic parameters to allow the following with min_const_generics:

impl Foo { fn test() -> [u8; std::mem::size_of::<Self>()] { todo!() } }

struct Bar([u8; baz::<Self>()]);
const fn baz<T>() -> usize { 10 }

We do however allow Self in repeat expression even if it is generic to not break code which already works on stable while causing the const_evaluatable_unchecked future compat lint:

fn foo<T>() {
    let _bar = [1_u8; std::mem::size_of::<*mut T>()];
}

is_trait_impl: bool

Is this within an impl Foo for bar?

The Self type, as used somewhere other than within a trait.

Belongs to the type namespace.

Examples:

struct Bar(Box<Self>); // SelfTyAlias

trait Foo {
    fn foo() -> Box<Self>; // SelfTyParam
}

impl Bar {
    fn blah() {
        let _: Self; // SelfTyAlias
    }
}

impl Foo for Bar {
    fn foo() -> Box<Self> { // SelfTyAlias
        let _: Self;        // SelfTyAlias

        todo!()
    }
}

See also Res::SelfCtor.

SelfCtor(DefId)

The Self constructor, along with the DefId of the impl it is associated with.

Belongs to the value namespace.

See also Res::SelfTyParam and Res::SelfTyAlias.

Local(Id)

A local variable or function parameter.

Belongs to the value namespace.

ToolMod

A tool attribute module; e.g., the rustfmt in #[rustfmt::skip].

Belongs to the type namespace.

NonMacroAttr(NonMacroAttrKind)

An attribute that is not implemented via macro. E.g., #[inline] and #[rustfmt::skip], which are essentially directives, as opposed to #[test], which is a builtin macro.

Belongs to the macro namespace.

Err

Name resolution failed. We use a dummy Res variant so later phases of the compiler won’t crash and can instead report more errors.

Not bound to a specific namespace.

Implementations

impl<Id> Res<Id>

pub fn def_id(&self) -> DefIdwhere Id: Debug,

Return the DefId of this Def if it has an ID, else panic.

pub fn opt_def_id(&self) -> Option<DefId>

Return Some(..) with the DefId of this Res if it has a ID, else None.

pub fn mod_def_id(&self) -> Option<DefId>

Return the DefId of this Res if it represents a module.

pub fn descr(&self) -> &'static str

A human readable name for the res kind (“function”, “module”, etc.).

pub fn article(&self) -> &'static str

Gets an English article for the Res.

pub fn map_id<R>(self, map: impl FnMut(Id) -> R) -> Res<R>

pub fn apply_id<R, E>( self, map: impl FnMut(Id) -> Result<R, E> ) -> Result<Res<R>, E>

pub fn expect_non_local<OtherId>(self) -> Res<OtherId>

pub fn macro_kind(self) -> Option<MacroKind>

pub fn ns(&self) -> Option<Namespace>

Returns None if this is Res::Err

pub fn matches_ns(&self, ns: Namespace) -> bool

Always returns true if self is Res::Err

pub fn expected_in_tuple_struct_pat(&self) -> bool

Returns whether such a resolved path can occur in a tuple struct/variant pattern

pub fn expected_in_unit_struct_pat(&self) -> bool

Returns whether such a resolved path can occur in a unit struct/variant pattern