Enum rustc_hir::Node

pub enum Node<'hir> {

    Param(&'hir Param<'hir>),
    Item(&'hir Item<'hir>),
    ForeignItem(&'hir ForeignItem<'hir>),
    TraitItem(&'hir TraitItem<'hir>),
    ImplItem(&'hir ImplItem<'hir>),
    Variant(&'hir Variant<'hir>),
    Field(&'hir FieldDef<'hir>),
    AnonConst(&'hir AnonConst),
    ConstBlock(&'hir ConstBlock),
    Expr(&'hir Expr<'hir>),
    ExprField(&'hir ExprField<'hir>),
    Stmt(&'hir Stmt<'hir>),
    PathSegment(&'hir PathSegment<'hir>),
    Ty(&'hir Ty<'hir>),
    TypeBinding(&'hir TypeBinding<'hir>),
    TraitRef(&'hir TraitRef<'hir>),
    Pat(&'hir Pat<'hir>),
    PatField(&'hir PatField<'hir>),
    Arm(&'hir Arm<'hir>),
    Block(&'hir Block<'hir>),
    Local(&'hir Local<'hir>),
    Ctor(&'hir VariantData<'hir>),
    Lifetime(&'hir Lifetime),
    GenericParam(&'hir GenericParam<'hir>),
    Crate(&'hir Mod<'hir>),
    Infer(&'hir InferArg),
    WhereBoundPredicate(&'hir WhereBoundPredicate<'hir>),
    ArrayLenInfer(&'hir InferArg),
    AssocOpaqueTy(&'hir AssocOpaqueTy),
    Err(&'hir Span),

}

Variants

Param(&'hir Param<'hir>)

Item(&'hir Item<'hir>)

ForeignItem(&'hir ForeignItem<'hir>)

TraitItem(&'hir TraitItem<'hir>)

ImplItem(&'hir ImplItem<'hir>)

Variant(&'hir Variant<'hir>)

Field(&'hir FieldDef<'hir>)

AnonConst(&'hir AnonConst)

ConstBlock(&'hir ConstBlock)

Expr(&'hir Expr<'hir>)

ExprField(&'hir ExprField<'hir>)

Stmt(&'hir Stmt<'hir>)

PathSegment(&'hir PathSegment<'hir>)

Ty(&'hir Ty<'hir>)

TypeBinding(&'hir TypeBinding<'hir>)

TraitRef(&'hir TraitRef<'hir>)

Pat(&'hir Pat<'hir>)

PatField(&'hir PatField<'hir>)

Arm(&'hir Arm<'hir>)

Block(&'hir Block<'hir>)

Local(&'hir Local<'hir>)

Ctor(&'hir VariantData<'hir>)

Ctor refers to the constructor of an enum variant or struct. Only tuple or unit variants with synthesized constructors.

Lifetime(&'hir Lifetime)

GenericParam(&'hir GenericParam<'hir>)

Crate(&'hir Mod<'hir>)

Infer(&'hir InferArg)

WhereBoundPredicate(&'hir WhereBoundPredicate<'hir>)

ArrayLenInfer(&'hir InferArg)

AssocOpaqueTy(&'hir AssocOpaqueTy)

Err(&'hir Span)

Implementations

impl<'hir> Node<'hir>

pub fn ident(&self) -> Option<Ident>

Get the identifier of this Node, if applicable.

Edge cases

Calling .ident() on a Node::Ctor will return None because Ctors do not have identifiers themselves. Instead, call .ident() on the parent struct/variant, like so:

ctor
    .ctor_hir_id()
    .map(|ctor_id| tcx.parent_hir_node(ctor_id))
    .and_then(|parent| parent.ident())

pub fn fn_decl(self) -> Option<&'hir FnDecl<'hir>>

pub fn fn_sig(self) -> Option<&'hir FnSig<'hir>>

pub fn ty(self) -> Option<&'hir Ty<'hir>>

Get the type for constants, assoc types, type aliases and statics.

pub fn alias_ty(self) -> Option<&'hir Ty<'hir>>

pub fn body_id(&self) -> Option<BodyId>

pub fn generics(self) -> Option<&'hir Generics<'hir>>

pub fn as_owner(self) -> Option<OwnerNode<'hir>>

pub fn fn_kind(self) -> Option<FnKind<'hir>>

pub fn expect_param(&self) -> &'hir Param<'hir>

pub fn expect_item(&self) -> &'hir Item<'hir>

pub fn expect_foreign_item(&self) -> &'hir ForeignItem<'hir>

pub fn expect_trait_item(&self) -> &'hir TraitItem<'hir>

pub fn expect_impl_item(&self) -> &'hir ImplItem<'hir>

pub fn expect_variant(&self) -> &'hir Variant<'hir>

pub fn expect_field(&self) -> &'hir FieldDef<'hir>

pub fn expect_anon_const(&self) -> &'hir AnonConst

pub fn expect_inline_const(&self) -> &'hir ConstBlock

pub fn expect_expr(&self) -> &'hir Expr<'hir>

pub fn expect_expr_field(&self) -> &'hir ExprField<'hir>

pub fn expect_stmt(&self) -> &'hir Stmt<'hir>

pub fn expect_path_segment(&self) -> &'hir PathSegment<'hir>

pub fn expect_ty(&self) -> &'hir Ty<'hir>

pub fn expect_type_binding(&self) -> &'hir TypeBinding<'hir>

pub fn expect_trait_ref(&self) -> &'hir TraitRef<'hir>

pub fn expect_pat(&self) -> &'hir Pat<'hir>

pub fn expect_pat_field(&self) -> &'hir PatField<'hir>

pub fn expect_arm(&self) -> &'hir Arm<'hir>

pub fn expect_block(&self) -> &'hir Block<'hir>

pub fn expect_local(&self) -> &'hir Local<'hir>

pub fn expect_ctor(&self) -> &'hir VariantData<'hir>

pub fn expect_lifetime(&self) -> &'hir Lifetime

pub fn expect_generic_param(&self) -> &'hir GenericParam<'hir>

pub fn expect_crate(&self) -> &'hir Mod<'hir>

pub fn expect_infer(&self) -> &'hir InferArg

pub fn expect_closure(&self) -> &'hir Closure<'hir>