rustc_hir::lang_items
Enum LangItem

pub enum LangItem {

    Sized,
    Unsize,
    StructuralPeq,
    Copy,
    Clone,
    CloneFn,
    Sync,
    DiscriminantKind,
    Discriminant,
    PointeeTrait,
    Metadata,
    DynMetadata,
    Freeze,
    FnPtrTrait,
    FnPtrAddr,
    Drop,
    Destruct,
    AsyncDrop,
    AsyncDestruct,
    AsyncDropInPlace,
    SurfaceAsyncDropInPlace,
    AsyncDropSurfaceDropInPlace,
    AsyncDropSlice,
    AsyncDropChain,
    AsyncDropNoop,
    AsyncDropDeferredDropInPlace,
    AsyncDropFuse,
    AsyncDropDefer,
    AsyncDropEither,
    CoerceUnsized,
    DispatchFromDyn,
    TransmuteOpts,
    TransmuteTrait,
    Add,
    Sub,
    Mul,
    Div,
    Rem,
    Neg,
    Not,
    BitXor,
    BitAnd,
    BitOr,
    Shl,
    Shr,
    AddAssign,
    SubAssign,
    MulAssign,
    DivAssign,
    RemAssign,
    BitXorAssign,
    BitAndAssign,
    BitOrAssign,
    ShlAssign,
    ShrAssign,
    Index,
    IndexMut,
    UnsafeCell,
    VaList,
    Deref,
    DerefMut,
    DerefPure,
    DerefTarget,
    Receiver,
    Fn,
    FnMut,
    FnOnce,
    AsyncFn,
    AsyncFnMut,
    AsyncFnOnce,
    AsyncFnOnceOutput,
    CallOnceFuture,
    CallRefFuture,
    AsyncFnKindHelper,
    AsyncFnKindUpvars,
    FnOnceOutput,
    Iterator,
    FusedIterator,
    Future,
    FutureOutput,
    AsyncIterator,
    CoroutineState,
    Coroutine,
    CoroutineReturn,
    CoroutineYield,
    CoroutineResume,
    Unpin,
    Pin,
    OrderingEnum,
    PartialEq,
    PartialOrd,
    CVoid,
    Panic,
    PanicNounwind,
    PanicFmt,
    ConstPanicFmt,
    PanicBoundsCheck,
    PanicMisalignedPointerDereference,
    PanicInfo,
    PanicLocation,
    PanicImpl,
    PanicCannotUnwind,
    PanicInCleanup,
    PanicAddOverflow,
    PanicSubOverflow,
    PanicMulOverflow,
    PanicDivOverflow,
    PanicRemOverflow,
    PanicNegOverflow,
    PanicShrOverflow,
    PanicShlOverflow,
    PanicDivZero,
    PanicRemZero,
    PanicCoroutineResumed,
    PanicAsyncFnResumed,
    PanicAsyncGenFnResumed,
    PanicGenFnNone,
    PanicCoroutineResumedPanic,
    PanicAsyncFnResumedPanic,
    PanicAsyncGenFnResumedPanic,
    PanicGenFnNonePanic,
    BeginPanic,
    FormatAlignment,
    FormatArgument,
    FormatArguments,
    FormatCount,
    FormatPlaceholder,
    FormatUnsafeArg,
    ExchangeMalloc,
    DropInPlace,
    FallbackSurfaceDrop,
    AllocLayout,
    Start,
    EhPersonality,
    EhCatchTypeinfo,
    OwnedBox,
    GlobalAlloc,
    PtrUnique,
    PhantomData,
    ManuallyDrop,
    MaybeUninit,
    AlignOffset,
    Termination,
    Try,
    Tuple,
    SliceLen,
    TryTraitFromResidual,
    TryTraitFromOutput,
    TryTraitBranch,
    TryTraitFromYeet,
    PointerLike,
    ConstParamTy,
    UnsizedConstParamTy,
    Poll,
    PollReady,
    PollPending,
    AsyncGenReady,
    AsyncGenPending,
    AsyncGenFinished,
    ResumeTy,
    GetContext,
    Context,
    FuturePoll,
    AsyncIteratorPollNext,
    IntoAsyncIterIntoIter,
    Option,
    OptionSome,
    OptionNone,
    ResultOk,
    ResultErr,
    ControlFlowContinue,
    ControlFlowBreak,
    IntoFutureIntoFuture,
    IntoIterIntoIter,
    IteratorNext,
    PinNewUnchecked,
    RangeFrom,
    RangeFull,
    RangeInclusiveStruct,
    RangeInclusiveNew,
    Range,
    RangeToInclusive,
    RangeTo,
    String,
    CStr,
    EffectsRuntime,
    EffectsNoRuntime,
    EffectsMaybe,
    EffectsIntersection,
    EffectsIntersectionOutput,
    EffectsCompat,
    EffectsTyCompat,

}

A representation of all the valid lang items in Rust.

Variants

Sized

The sized lang item.

Unsize

The unsize lang item.

StructuralPeq

The structural_peq lang item. Trait injected by #[derive(PartialEq)], (i.e. “Partial EQ”).

Copy

The copy lang item.

Clone

The clone lang item.

CloneFn

The clone_fn lang item.

Sync

The sync lang item.

DiscriminantKind

The discriminant_kind lang item.

Discriminant

The discriminant_type lang item. The associated item of the DiscriminantKind trait.

PointeeTrait

The pointee_trait lang item.

Metadata

The metadata_type lang item.

DynMetadata

The dyn_metadata lang item.

Freeze

The freeze lang item.

FnPtrTrait

The fn_ptr_trait lang item.

FnPtrAddr

The fn_ptr_addr lang item.

Drop

The drop lang item.

Destruct

The destruct lang item.

AsyncDrop

The async_drop lang item.

AsyncDestruct

The async_destruct lang item.

AsyncDropInPlace

The async_drop_in_place lang item.

SurfaceAsyncDropInPlace

The surface_async_drop_in_place lang item.

AsyncDropSurfaceDropInPlace

The async_drop_surface_drop_in_place lang item.

AsyncDropSlice

The async_drop_slice lang item.

AsyncDropChain

The async_drop_chain lang item.

AsyncDropNoop

The async_drop_noop lang item.

AsyncDropDeferredDropInPlace

The async_drop_deferred_drop_in_place lang item.

AsyncDropFuse

The async_drop_fuse lang item.

AsyncDropDefer

The async_drop_defer lang item.

AsyncDropEither

The async_drop_either lang item.

CoerceUnsized

The coerce_unsized lang item.

DispatchFromDyn

The dispatch_from_dyn lang item.

TransmuteOpts

The transmute_opts lang item.

TransmuteTrait

The transmute_trait lang item.

Add

The add lang item.

Sub

The sub lang item.

Mul

The mul lang item.

Div

The div lang item.

Rem

The rem lang item.

Neg

The neg lang item.

Not

The not lang item.

BitXor

The bitxor lang item.

BitAnd

The bitand lang item.

BitOr

The bitor lang item.

Shl

The shl lang item.

Shr

The shr lang item.

AddAssign

The add_assign lang item.

SubAssign

The sub_assign lang item.

MulAssign

The mul_assign lang item.

DivAssign

The div_assign lang item.

RemAssign

The rem_assign lang item.

BitXorAssign

The bitxor_assign lang item.

BitAndAssign

The bitand_assign lang item.

BitOrAssign

The bitor_assign lang item.

ShlAssign

The shl_assign lang item.

ShrAssign

The shr_assign lang item.

Index

The index lang item.

IndexMut

The index_mut lang item.

UnsafeCell

The unsafe_cell lang item.

VaList

The va_list lang item.

Deref

The deref lang item.

DerefMut

The deref_mut lang item.

DerefPure

The deref_pure lang item.

DerefTarget

The deref_target lang item.

Receiver

The receiver lang item.

Fn

The Fn lang item.

FnMut

The fn_mut lang item.

FnOnce

The fn_once lang item.

AsyncFn

The async_fn lang item.

AsyncFnMut

The async_fn_mut lang item.

AsyncFnOnce

The async_fn_once lang item.

AsyncFnOnceOutput

The async_fn_once_output lang item.

CallOnceFuture

The call_once_future lang item.

CallRefFuture

The call_ref_future lang item.

AsyncFnKindHelper

The async_fn_kind_helper lang item.

AsyncFnKindUpvars

The async_fn_kind_upvars lang item.

FnOnceOutput

The fn_once_output lang item.

Iterator

The iterator lang item.

FusedIterator

The fused_iterator lang item.

Future

The future_trait lang item.

FutureOutput

The future_output lang item.

AsyncIterator

The async_iterator lang item.

CoroutineState

The coroutine_state lang item.

Coroutine

The coroutine lang item.

CoroutineReturn

The coroutine_return lang item.

CoroutineYield

The coroutine_yield lang item.

CoroutineResume

The coroutine_resume lang item.

Unpin

The unpin lang item.

Pin

The pin lang item.

OrderingEnum

The Ordering lang item.

PartialEq

The eq lang item.

PartialOrd

The partial_ord lang item.

CVoid

The c_void lang item.

Panic

The panic lang item.

PanicNounwind

The panic_nounwind lang item.

PanicFmt

The panic_fmt lang item.

ConstPanicFmt

The const_panic_fmt lang item.

PanicBoundsCheck

The panic_bounds_check lang item.

PanicMisalignedPointerDereference

The panic_misaligned_pointer_dereference lang item.

PanicInfo

The panic_info lang item.

PanicLocation

The panic_location lang item.

PanicImpl

The panic_impl lang item.

PanicCannotUnwind

The panic_cannot_unwind lang item.

PanicInCleanup

The panic_in_cleanup lang item.

PanicAddOverflow

The panic_const_add_overflow lang item. Constant panic messages, used for codegen of MIR asserts.

PanicSubOverflow

The panic_const_sub_overflow lang item.

PanicMulOverflow

The panic_const_mul_overflow lang item.

PanicDivOverflow

The panic_const_div_overflow lang item.

PanicRemOverflow

The panic_const_rem_overflow lang item.

PanicNegOverflow

The panic_const_neg_overflow lang item.

PanicShrOverflow

The panic_const_shr_overflow lang item.

PanicShlOverflow

The panic_const_shl_overflow lang item.

PanicDivZero

The panic_const_div_by_zero lang item.

PanicRemZero

The panic_const_rem_by_zero lang item.

PanicCoroutineResumed

The panic_const_coroutine_resumed lang item.

PanicAsyncFnResumed

The panic_const_async_fn_resumed lang item.

PanicAsyncGenFnResumed

The panic_const_async_gen_fn_resumed lang item.

PanicGenFnNone

The panic_const_gen_fn_none lang item.

PanicCoroutineResumedPanic

The panic_const_coroutine_resumed_panic lang item.

PanicAsyncFnResumedPanic

The panic_const_async_fn_resumed_panic lang item.

PanicAsyncGenFnResumedPanic

The panic_const_async_gen_fn_resumed_panic lang item.

PanicGenFnNonePanic

The panic_const_gen_fn_none_panic lang item.

BeginPanic

The begin_panic lang item. libstd panic entry point. Necessary for const eval to be able to catch it

FormatAlignment

The format_alignment lang item.

FormatArgument

The format_argument lang item.

FormatArguments

The format_arguments lang item.

FormatCount

The format_count lang item.

FormatPlaceholder

The format_placeholder lang item.

FormatUnsafeArg

The format_unsafe_arg lang item.

ExchangeMalloc

The exchange_malloc lang item.

DropInPlace

The drop_in_place lang item.

FallbackSurfaceDrop

The fallback_surface_drop lang item.

AllocLayout

The alloc_layout lang item.

Start

The start lang item.

EhPersonality

The eh_personality lang item.

EhCatchTypeinfo

The eh_catch_typeinfo lang item.

OwnedBox

The owned_box lang item.

GlobalAlloc

The global_alloc_ty lang item.

PtrUnique

The ptr_unique lang item.

PhantomData

The phantom_data lang item.

ManuallyDrop

The manually_drop lang item.

MaybeUninit

The maybe_uninit lang item.

AlignOffset

The align_offset lang item. Align offset for stride != 1; must not panic.

Termination

The termination lang item.

Try

The Try lang item.

Tuple

The tuple_trait lang item.

SliceLen

The slice_len_fn lang item.

TryTraitFromResidual

The from_residual lang item.

TryTraitFromOutput

The from_output lang item.

TryTraitBranch

The branch lang item.

TryTraitFromYeet

The from_yeet lang item.

PointerLike

The pointer_like lang item.

ConstParamTy

The const_param_ty lang item.

UnsizedConstParamTy

The unsized_const_param_ty lang item.

Poll

The Poll lang item.

PollReady

The Ready lang item.

PollPending

The Pending lang item.

AsyncGenReady

The AsyncGenReady lang item.

AsyncGenPending

The AsyncGenPending lang item.

AsyncGenFinished

The AsyncGenFinished lang item.

ResumeTy

The ResumeTy lang item.

GetContext

The get_context lang item.

Context

The Context lang item.

FuturePoll

The poll lang item.

AsyncIteratorPollNext

The async_iterator_poll_next lang item.

IntoAsyncIterIntoIter

The into_async_iter_into_iter lang item.

Option

The Option lang item.

OptionSome

The Some lang item.

OptionNone

The None lang item.

ResultOk

The Ok lang item.

ResultErr

The Err lang item.

ControlFlowContinue

The Continue lang item.

ControlFlowBreak

The Break lang item.

IntoFutureIntoFuture

The into_future lang item.

IntoIterIntoIter

The into_iter lang item.

IteratorNext

The next lang item.

PinNewUnchecked

The new_unchecked lang item.

RangeFrom

The RangeFrom lang item.

RangeFull

The RangeFull lang item.

RangeInclusiveStruct

The RangeInclusive lang item.

RangeInclusiveNew

The range_inclusive_new lang item.

Range

The Range lang item.

RangeToInclusive

The RangeToInclusive lang item.

RangeTo

The RangeTo lang item.

String

The String lang item.

CStr

The CStr lang item.

EffectsRuntime

The EffectsRuntime lang item.

EffectsNoRuntime

The EffectsNoRuntime lang item.

EffectsMaybe

The EffectsMaybe lang item.

EffectsIntersection

The EffectsIntersection lang item.

EffectsIntersectionOutput

The EffectsIntersectionOutput lang item.

EffectsCompat

The EffectsCompat lang item.

EffectsTyCompat

The EffectsTyCompat lang item.

Implementations

impl LangItem

fn from_u32(u: u32) -> Option<LangItem>

pub fn name(self) -> Symbol

Returns the name symbol in #[lang = "$name"]. For example, LangItem::PartialEq.name() would result in sym::eq since it is #[lang = "eq"].

pub fn from_name(name: Symbol) -> Option<Self>

Opposite of LangItem::name

pub fn variant_name(self) -> &'static str

Returns the name of the LangItem enum variant.

pub fn target(self) -> Target

pub fn required_generics(&self) -> GenericRequirement

impl LangItem

pub fn is_weak(self) -> bool

pub fn link_name(self) -> Option<Symbol>