rustc_hir
Trait HashStableContext

pub trait HashStableContext: HashStableContext + HashStableContext { }

Requirements for a StableHashingContext to be used in this crate. This is a hack to allow using the HashStable_Generic derive macro instead of implementing everything in rustc_middle.